refactor: 엔티티 패키지 수정
