+++ src/main/java/com/example/emotion_storage/user/SocialType.java
@@ -1,4 +1,4 @@
-package com.example.emotion_storage.entity;
+package com.example.emotion_storage.user;
 
 public enum SocialType {
     GOOGLE, KAKAO
